+++ openssl/src/x509/tests.rs
@@ -266,6 +266,39 @@ fn test_subject_alt_name_iter() {
 }
 
 #[test]
+fn test_name_constraints() {
+    let cert = include_bytes!("../../test/name_constraints.pem");
+    let cert = X509::from_pem(cert).unwrap();
+
+    // Access through .name_constraints()
+    let name_constraints = cert
+        .name_constraints()
+        .expect("Name constraints extension should be present");
+
+    let permitted = name_constraints.permitted().unwrap();
+    assert_eq!(permitted.len(), 1);
+
+    let base = permitted[0].base().dnsname().unwrap();
+    assert_eq!(base, ".example.com");
+
+    let minimum = permitted[0].minimum().is_none();
+    assert_eq!(minimum, true);
+    let maximum = permitted[0].maximum().is_none();
+    assert_eq!(maximum, true);
+
+    let excluded = name_constraints.excluded().unwrap();
+    assert_eq!(excluded.len(), 1);
+
+    let base = excluded[0].base().dnsname().unwrap();
+    assert_eq!(base, ".example.net");
+
+    let minimum = excluded[0].minimum().is_none();
+    assert_eq!(minimum, true);
+    let maximum = excluded[0].maximum().is_none();
+    assert_eq!(maximum, true);
+}
+
+#[test]
 fn test_aia_ca_issuer() {
     // With AIA
     let cert = include_bytes!("../../test/aia_test_cert.pem");